Making example slightly bigger.

+++ trunk/old.werkflow/werkflow-core/src/java/example/com/werken/werkflow/example/hanoi/Example.java
@@ -20,7 +20,7 @@ public class Example
         SimpleAttributes attrs = new SimpleAttributes();
 
         attrs.setAttribute( "height",
-                            new Integer( 5 ) );
+                            new Integer( 6 ) );
 
         getWfms().getRuntime().callProcess( "hanoi-solver",
                                             attrs );
